Audit + CleanUp
git-svn-id: https://cchlib.googlecode.com/svn/trunk@8247 bc589049-105d-1923-a3ce-6fcb24364ed1

+++ cchlib-core-sample/src/main/java/paper/reflexion/ChildClass.java
@@ -62,7 +62,7 @@ public class ChildClass extends LegacyParentClass
             }
 
         System.out.println( "***********" );
-        int aProtectedAbstractMethod = this.parentAcces.silentCall( this, Integer.class, "aProtectedAbstractMethod" );
+        int aProtectedAbstractMethod = this.parentAcces.silentCall( this, Integer.class, "aProtectedAbstractMethod" ).intValue();
         System.out.println( "aProtectedAbstractMethod()=" + aProtectedAbstractMethod );
 
         System.out.println( "***********" );
